requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
codegen==1.0
dj-database-url==0.5.0
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
mysqlclient==2.1.0
patterns==0.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
